Fake client test application
The test application 'client_mmap.c' is a simple demonstration of
message passing between two process.  The app allows for two methods of
data exchange:  one which uses a shared mmap file / buffer;  and another
which is based on pipes and vmsplice.  Both methods use a pipe for event
notification.  Currently, the application will memcpy to / from a local
buffer to produce a memory bandwidth profile similar to that of a niova
user process and the local niova on the same node.