Local Registry Subsystem
This patch adds a "local registry" which is basically a tree of nodes
which try to mimic JSON nodes.  The idea behind "lreg" is to provide a
subsystem which can be used to hold arbitrary objects in a namespace
which can be translated into KVs as needed.

The first users of the lreg are log entries.  The idea here is that each
log entry is made into a static object which is installed into a node
owned by the respective .c or .h file containing the log message.
Later, we will allow users to modify the log levels via external input
sources.