Add thread.[ch] / Reclassify lreg val / LREG list
This patch contains 3 logical changes (which should have been sent in
separate patches).

First, the logging subsystem was using thread IDs which weren't very
informational.  This patch employs pthread_setname_np() to address this.
The functions that had been in log.[ch] have been moved into a new set
of files spefically for holding thread-related functions.

Next, struct lreg_value has been rebuilt so that there are clear
delineation of 'put' and 'get' values.  This change was done after
considering the scalability of the registry and the unsavory fact that
having one lreg_node per attribute is not very practical.  With this
reworking of the lreg_value, the user can now pass in a sub-key for read
or write purposes and this key can be interpreted by the callback in a
more arbitrary manner.  This will allow a single object to provide
numerous k/v pairs.

The linkage of the lreg_node is now doubly linked.  The use of a
singly-linked list was a bit optimistic.  This change was prompted by
the desire to keep the registry as single-threaded as possible.
Therefore, the install function now just queues the installing lreg_node
onto a list and the lreg service thread will process that node at some
point.  Since this install queue may contain dependencies, using a
singly-linked list in the lreg_node would no longer work.

Finally, this patch contains an example of a gcc destructor being used
for the function lreg_subsystem_destroy().