Thread control and other cleanups
This patch adds thread.[ch] so that thread startup and shutdown can be
abstracted.  The existing threads in the code were adapted to the new
API.

The new thread ctl API currently includes:  thread_create, which applies
a name to the thread;  thread_run, thread_halt, and thread_pause which
allow for the threads run state to be modifed;  and a macro called,
THREAD_LOOP_WITH_CTL, which implements the main loop of the thread.

Moved some items from common.h to newly added header files, util.h and
ctor.h.