Allow log levels to be non-const
The static initialization of .lei_level didn't take kindly to being
init'd by a value that was resolved during run time.  This patch changes
the logic a bit to detect if the log_entry_info node was changed, and if
it was, only then is the stored log level taken.  Otherwise, the one
provided in the debug msg is used.