niosd_io_test improvements
The niosd_io_test is now starting to look like a more full-fledged
testing program for the niosd api.  The code has largely been rewritten
to accept all sort of new parameters.  In the process of making this
work, a bug was fixed in niosd_device_open() where the size of block
devices was not being measured correctly.