Add blocking support to niosd io layer
This patch introduces the ability for a thread to block on a pipe fd
to wait for event handler completion.

The niosd_ctx gets several new function calls to provide this feature.
To enable the capability, niosd_device_params_enable_blocking_mode() is
called with the niosd_io_ctx_type to which blocking support is to be
applied.  Internally, the niosd_device initialization will then call
nioctx_blocking_mode_setup() to create a pipe.  The read end of the pipe
may be obtained with the call, nioctx_blocking_mode_fd_get().  The pipe
is written by nioctx_blocking_mode_notify(), which is called from
niosd_device_event_thread_post_new_events() in event handler context.

The niosd_io test has been updated to allow for this blocking mode to be
employed.  However, the usage implemented in niosd_io_test is a bit
kludgy since the test doesn't run like an I/O service process and there
are few workarounds to contort to the blocking mode usage.  With this
being the case, the niosd_io_test does wrap the pipe fd within an epoll
fd instance and this method does appear to work.

In the future, I anticipate that the io-submitter thread will have other
responsibilities, such as blocking on the network.  This will require
epoll so that the niosd-dev pipe can be watched along with any other
sockets or file descriptors.