Add epoll management
Introduce a simple epoll API wrapper so that a callback can be provided
through a uniform epoll handle structure.

The '-E' mode in niosd_io_test has been modified to use this new API
(though the code is a bit of a mess admittedly).

Remove 'local_' from registry and ctl_interface

It didn't appear that the 'local_' prefix was adding much info.