Signal handling for graceful niosd_io thread exit
This patch also introduces the niova-ctl binary.  It was when testing
very basic operation of niova-ctl that I noticed the process was
segfaulting.  The segfault was due to io_destroy() being called while
the niosd_io_getevents thread was still blocked.  To deal with this, a
timeout was added to io_getevents(), but one that was still a bit long
for interactive use (10 seconds).  A signaling method was added to cause
the threads to return from io_getevents() with -EINTR.