Config token processing
This patch adds support for reading and parsing config files.  The
config file format is very simple and the terms 'token' and 'key' are
mostly interchangeable at this point since the current schema is
basically a K/V pair.

The conf file contents are simply one or more lines of the following:

TOKEN value
TOKEN value
TOKEN value
...

The config_token library defines a fixed set of tokens which users of
the library may select one or more tokens for their usage.

The main structure is 'conf_token_set_parser', which is a handle for
input + output buffers, buffer size values, and the set of tokens to be
applied to the parsing operation.  conf_token_set_parse() is called to
open, read, parse the file according to the set of enabled tokens.  For
each properly matched token, a user callback is invoked which provides
the token and value contents.  An optional regex may be applied to the
value to ensure that it meets the predefined specifications.  At this
time, we currently have regex's for UUID and IPADDR.

test/config_token_test.c is a simple example of usage.