Add ctl-svc bootstrap bits
This patch enables the processing of a ctl-svc "local" repo of config
information in the form on config-token files.  Currently, the processed
directory is:  /etc/niova/ctl-svc/local.  If this directory does not
exist then the processing is skipped.  If the directory is present then
the contents will be parsed of the file names which match the UUID regex
in regex-defines.h.  At this time, only parsing occurs, there's no
object creation which results from the presence of the config files.
Soon the ctl_svc_node structure will be used to establish a set of
entities which were found during the parsing of the ctl-svc dir.

To override the default ctl-svc dir, the enviroment variable
"NIOVA_LOCAL_CTL_SVC_DIR" may be set with an alternative directory.