Raft program adds udp sockets to epoll_mgr
The epoll_mgr in raft.c now includes the udp sockets.  Currently the
udp epoll callback, raft_server_udp_cb(), is basically a noop.

The udp.c api was changed a bit so that the socket() and bind() system
calls are separated into two different functions.  This is so that data
do not arrive on the socket before other preparations are done, such as
adding the socketfd to an epoll set.