IOV related calls to io.[ch]
io_iovs_total_size_get() helper function which tallies the total size of
an iov array.

io_iovs_map_consumed() is a call which creates a new iov array from an
existing one and the number of bytes which have already been consumed
from the existing iov.

Also added io_fd_drain(), which can be used for reading all of the data
from a file descriptor.  This can be used for event fd's.