Raft test implementation checkpoint
This is set of patches which have been squashed into a large commit.
The patch here advances the implementation of the raft engine quite a
bit but it's still not fully operational.  At this time, leadership
elections and append-entries heartbeat appear to be working somewhat.
Significant portions are still missing, however, mainly the support for
pushing out append-entries which have payload.  The code for doing log
rollback is somewhat in place and will need further work after
append-entries becomes more complete.