Raft restructuring and start of client code
This patch restructures the raft code a bit moving some portions between
raft_server and raft_net.  The raft instance now accepts custom timer_fd
and udp_recv callbacks to be used with epoll.  This will allow the test
client to operate in a similar process structure to that of the raft
server process.

At this time, the raft test client doesn't yet do anything.  The next
step will be for the client to send a msg which has to be propagated
through the append-entries RPC and stored to the members' raft logs.
Once this is shown to be working, and server's logs can be reconstructed
when found to be incomplete, then I'll start to implement a simple
client application which uses a deterministic stream of random numbers
to test the correction of the implementation.