Raft restructuring patch and some client mods
While looking into the raft_client_test code, I decided to refactor some
pieces to (hopefully) unify the overall raft code.

One large aspect of this patch is the shift of the udp and timerfd
callbacks to be more application specific.  There are now generic
callbacks for each which perform the I/O and then call the
application-specified callback.

Timespecs have been added to the raft instance so that communication
timestamps can be maintained for peers.  This will have some importance
for clients, which must decide on which server(s) to call.