Raft append entries request handler implementation
This patch attempts to complete the AE request handling support for our
raft engine.  There's quite a few important elements here in addition to
the explicit handling of the AE RPC.  One such element is log truncation
which is done by the function, raft_server_log_truncate().  Otherwise,
the patch tries to implement the AE procedure as explained in the Raft
paper.