Raft raft_server_becomes_follower() modifications
The function that was called, raft_server_candidate_becomes_follower()
is now called raft_server_becomes_follower().  This change was done so
that demotion to follower could be done from both the candidate state as
well as the leader state.

This patch also introduces an incomplete implementation of
raft_server_process_append_entries_reply() which uses
raft_server_becomes_follower().