Fix epoll mgr bug
Epoll mgr was calling epoll_wait with a number of events which was
greater than the size of the ev array.  This patch fixes that problem
and increases the preset sizes for epoll mgr events.  The reason for
this is because I believe that in the general case we should have one
event per fd in the set.