Raft checkpoint commit
For the last few days, I've been working on append entry implementation
and this has taken me through a bunch of different areas.  Here I'm
going to push the latest changes mainly for backup purposes.

Some changes here:

Add raerqm_leader_term and raerqm_log_term to replace raerqm_term in the
raft_append_entries_request_msg.  The reason for this is because old
entries being replayed to followers need the term in which they were
written, not the current leader's term, for proper replay detection.

raerpm_prev_log_index has been added to the raft_append_entries_reply_msg
so that the leader's next-idx[follower] is only incremented if the reply
corresponds to next-idx[follower] value.   This is to avoid incrementing
when the reply was delayed in the network.

Added ev_pipe's for append-entry-send and state-machine-apply
scheduling.  At this time the epoll handlers don't do anything but that
should change soon.

Added rls_ae_sends_wait_until to raft_leader_state.  This is to help
avoid aggressive AE retries to followers which are down.

Reworked the reh_index code so that reh_index reflects the raft index
NOT and the physical log index.  Log header blocks will now have a
negative reh_index value.

Did some work in an attempt to detect and properly handle retried AE
requests and replies.  raft_server_append_entry_check_already_stored()
has been added to check if the incoming AE corresponds to an already
stored entry.  The is meant to handle the case where msgs are delayed in
the network and/or replies to the leader are dropped.  The raft spec
dictates that this retry case is handled properly.  The code here will
only check for 'already_stored' if the follower's raft_entry_index is
greater than the one specified in the AE request.  At this time,
raft_server_append_entry_check_already_stored() checks the index
and term of the stored index against the RPC, as well as the prev-term.
At some point it may be worth sending a CRC in the request so that the
contents can be checked as well.

I'm sure there are tons of bugs at this point!