raft_server_state_machine_apply() implementation
This patch creates a very basic version of
raft_server_state_machine_apply().  If all goes well (which it likely
will not since this raft code hasn't been tried in a while), this patch
should allow for commit and application of the leader_change_marker noop
needed for a leader to clear one of the conditions for accepting client
application reads.