Raft reporting enhancements / bug fix
Fix off-by-one in raft_instance_lreg_cb().

Add last-applied-cumulative-crc to the server and the monitoring
interface.  This accumulates all of the crc's of the applied entries
into a single value.

Unfortunately, JSON does not support hexadecimal so I've converted
several log entries to use decimal so that they match the monitoring
output.