Raft followers are tolerant to negative log term
There is a case where the leader will send a log-term value of -1.  This
happens for heartbeat msgs issued when the leader and follower were in
sync.  However, if this msg is delayed and recv'd by the follower after
its log has advanced, the function
raft_server_append_entry_check_already_stored() will be called to test
if the entry has already been recv'd.  Previously,
raft_server_append_entry_check_already_stored() was asserting that the
log-term value was > 0.  This patch adjusts the code so that only the
previous block is verified in this situation.

There are lots of log level adjustments in this patch as well.