Raft client work
This patch starts to reimplement a bunch of internals to prepare for
client interaction.  The primary change here is a simplification of the
state machine callback parameters and cb functions - there is now only
one cb which takes a single parameter.  This new parameter, struct
raft_net_client_request, acts as both an input and output to the state
machine layer.