Raft server state machine implementation
This patch attempts to complete the server-side implementation for the
raft reference client app.  The test application has been modified
(though it wasn't operational or complete prior) to simplify its RPC
types, of  which is there is now just one.

struct raft_net_client_request, which is the bridge between the state
machine and the raft core, has had some items added to deal with stale
leaders and the source of the msg - which in the case of commit, is from
the raft log (not an incoming RPC).

raft_net_server_instance_run() has been modified to accept a single
state machine callback which is to handle the read, write, and commit
cases.

raft_server_send_msg_to_client() was simplified to use a only a single
source buffer which allowed the removal of the second IOV.

raft_server_udp_client_reply_init() does some basic setup on the reply
RPC so that the state-machine layer doesn't need to.

The remaining items in this patch, which are the majority, are in
raft_server_test.c.  Prior to this change, raft_server_test.c had no
actual state machine handling whatsover.  This patch introduces a ref
tree of struct rst_sm_node's each of which hold the state machine info
for a given "application instance".  Each application has a UUID.

The application, which is not yet complete, will allow for extensive
raft testing by storing and managing a set of computable sequences, one
per application instance.  The idea is to have the client verify the
correctness of the entire sequence by performing query and store
operations to the raft service.

The remainder of the patch attempts to deal with error handling in the
application (such as detecting > 1 operation in flight, invalid write
sequences, malformed RPCs, etc.) as well as the server side operation.
Errors here would be ones such as memory allocation failure or use of
incorrectly sized src / sink buffers.