Raft client implementation progress
Raft client test now deals with ping, ping replies and is close to
supporting read and writes.  Reads likely work but currently the write
path is not operational so a read cannot be tested.

This patch implements a lot of client-side mechanics for determining
which operations can be run based on the perceived availability of the
raft service.  The implementations for read and write support are largely
in place but still need debugging.