Add 'local' mode to random test
Local mode uses a privately allocated random buffer.  This is meant to
recreate the usage of the raft client.  I did try briefy to use
setstate_r() to replicate the contents of one generator instance and
pass it to another, however, I had mixed results.  In some select cases
I was able to end up with 2 generators producing the same stream of
randoms but this depended on the iteration number in which the
replication was done.