New lreg macros and log_subsystem registry entry
This patch adds a new registry object, log_subsystem.  It looks like:

	"log_subsystem" : {
                "master_log_level" : "debug",
                "log_msg_date_format" : "%blah"
        },

Currently, it has the ability to change the log level using the APPLY /
WHERE method.  The 'log_msg_date_format' stuff isn't completed yet.

The 'log_subsystem' object is presented as a direct child of the root
entry which is atypical since direct children are usually arrays.
However, it's become clear in a few cases where objects would be
preferred to arrays (as in this case).  A new macro has been introduced,
LREG_ROOT_ENTRY_GENERATE_OBJECT, which is similar to
LREG_ROOT_ENTRY_GENERATE except that it auto-creates a child object
which takes a multi-facet cb.  Overall, this saves a lot of boilerplate
code and I'll look to apply this method elsewhere.