Auto create the ctl-interface path based on UUID
If the UUID has been auto-detected then build the ctl-interface path
based on that UUID.  If the UUID has not be found, then the pid may be
used as a unique target for the NIOVA process.

In addition, a new env var has been introduced, NIOVA_INOTIFY_BASE_PATH.
This env var allows the user to specify a base path, rather than a full
path such as NIOVA_INOTIFY_PATH.  The difference is that the base path
option will apply the UUID / pid sub-directory whereas
NIOVA_INOTIFY_PATH will not.