ctl_svc_process_conf_file() fix for large files
The program was asserting when the file size was too large because the
assert was comparing a signed and unsigned value.

ctl_svc_init_scan_entries() will not call
ctl_svc_init_scan_entries_dump_invalid_file() unless a non-zero line is
present.  A line number of 0 indicates that the error occurred prior to
parsing.