Initial pass at fault injection feature
Fault injection is added in this patch.  The implementation uses a
hybrid static / global approach so that placeholders are never needed in
the case where an injection point is requested by a user prior to its
self-installation.  All injections are declared in a table,
faultInjections, and self-installation involves a call to
fault_injection_lookup() with the ID of the fault.  The IDs should be
unique to each fault injection point (though it is technically possible to
overload them per this implementation).

The subsystem is enabled when the code is compiled with the flag,
NIOVA_FAULT_INJECTION_ENABLED.

At this time, the "enabled" key is writable through the APPLY / WHERE
mechanism.

Here are the 2 example fault injections points contained in this patch:

        "fault_injection_points" : [
                {
                        "name" : "disabled injection",
                        "enabled" : false,
                        "file" : "src/raft_server.c",
                        "function" : "raft_server_main_loop",
                        "line_number" : 3562,
                        "when_to_inject" : "one_time_only",
                        "last_injected_at" : "Wed Dec 31 19:00:00 1969",
                        "injection_count" : 0,
                        "frequency_seconds" : 0,
                        "num_remaining" : 0,
                        "cond_exec_count" : 6
                },
                {
                        "name" : "any injection",
                        "enabled" : true,
                        "file" : "src/raft_server.c",
                        "function" : "raft_server_main_loop",
                        "line_number" : 3563,
                        "when_to_inject" : "every_time",
                        "last_injected_at" : "Tue Apr 28 14:02:59 2020",
                        "injection_count" : 6,
                        "frequency_seconds" : 0,
                        "num_remaining" : 0,
                        "cond_exec_count" : 6
                }
        ]
}