Ctl-interface cmd helper script
ctl-interface-cmd-to-all.sh copies a command file into the ctl-interface
directories which are named as UUIDs.  The script has an option to
iterate 'i' times each second and the ability to launch terminals to
edit the cmd file and/or watch the output.

Usage:  scripts/ctl-interface-cmd-to-all.sh [-d ctl-dir (/tmp/.niova/)]
	[-i iterations (1)] [-e (launch edit terminal)]
	[-w (launch watch terminal)] <cmd-file>