Add write-batch option to rocksdb test
This patch adds the option of doing rocksdb_writebatch_put() and
rocksdb_write() instead of rocksdb_put().  The write batch allows for
multiple K/V updates in an atomic update.  Here, the test program
updates the same key, HEADER-<uuid>, each time a new key is added.  This
blatant per-write overwriting is intentional so that the overhead of
this type of operation can be measured.