Add iterator operations to rocksdb test
This patch adds demonstration of the basic iterator operations in
rocksdb.  Seek, next, and prev operations are used to show that ordering
semantics are maintained in the key space.  Header and "last-key" KVs
are put into the db to act as location points where, respectively, the
first and last "application" entries may be found.