Basic ctl-interface endpoint monitoring
This patch enables the lookout to do some very basic monitoring
operations.  The results are reporting through http to the caller.
The reported results are intentionally verbose, the http interface is
meant to provide detailed info about the niova services on the node.

Note there's still a lot of work to do.. For one, the cmd file inputs and
outputs are not being cleaned up.  Also, dead enpoints are not yet
removed from the map (perhaps they will remain there forever but with
a lower polling frequency?).  Another item is replacing the UUID (used
as a key below) with a short name like "raft-0" or "nosd-1".

Example output:

$ curl -s http://localhost:8080/v0/
{
  "bf69c910-a426-11ea-afe6-90324b2d1e89": {
    "name": "r-a4e1",
    "type": "raft",
    "port": 6666,
    "responsive": true,
    "ep_info": {
       "system_info": {
          "current_time": {
               "time": "2020-06-01T22:05:09Z"
          },
          "start_time": {
               "time": "2020-06-01T16:41:32Z"
          },
          "pid": 10173,
          "uuid": "bf69c910-a426-11ea-afe6-90324b2d1e89",
          "ctl_interface_path": "//tmp/.niova/bf69c910-a426-11ea-afe6-90324b2d1e89/",
          "command_line": "/home/pauln/Code/niovad/raft-server -R -r bf692320-a426-11ea-b661-90324b2d1e89 -u bf69c910-a426-11ea-afe6-90324b2d1e89",
          "uts.nodename": "localhost.localdomain",
          "uts.sysname": "Linux",
          "uts.release": "5.2.17-100.fc29.x86_64",
          "uts.version": "#1 SMP Mon Sep 23 14:09:29 UTC 2019",
          "uts.machine": "x86_64",
          "rusage.user_cpu_time_used": 86.786644,
          "rusage.system_cpu_time_used": 22.139437,
          "rusage.max_rss": 510516,
          "rusage.min_fault": 155418,
          "rusage.maj_fault": 66,
          "rusage.in_block": 17192,
          "rusage.out_block": 73592,
          "rusage.vol_ctsw": 79335,
          "rusage.invol_ctsw": 10392
       }
     }
   }
}