Add raft reporting and cmd line arg support
Added structures to support the unmarshalling of raft info.

Now have a basic cmdline arg parser and a few params:

$ ./niova-lookout -h
Usage: [OPTIONS] ./niova-lookout
  -dir string
    	endpoint directory root (default "/tmp/.niova")
  -h
  -help
    	print help
  -port int
    	http listen port (default 8080)

This patch has some changes relating to the usage of sleep and high CPU
usage.  Today, with just 4 endpoints, I observed CPU usage around 15%.
This is extremely high for what's being done in the program.  I was
already planning to use inotify to watch the endpoint output directories
and this should alleviate the need for sleeping via time.Sleep().