PumiceDB lookup and write partial implementations
This patch mostly deals with the implementation of pmdb_object_lookup()
and its supporting code.  pmdb_object_lookup() starts to work with the
rocksDB api to lookup the KV for a PMDB object.  A PMDB object will have
binary contents in the form of struct pmdb_object, which has been
defined in this patch.  Note that pmdb_object_lookup() is not concerned
with the application API at all.  It's only used within the PumiceDB
private column-family for accessing the shadow metadata which PumiceDB
stores on behalf of all application objects.

The patch also implements a method for determining whether a write
operation is pending to application object through the use of a term
value held in the pmdb_object.  A term which is == to the current term
means that the a write is pending and no others may be accepted until it
completes.  It also means that the apply of the object (through commit)
must reset that term value.  The pmdb_object::term value is neutralized
when a leader election occurs and the held term is < the current raft
term.

pmdb_sm_handler_client_write() is also coming along but still needs some
work.  It also uses the newly introduced lookup method to obtain the
application's pumice metadata.  The error handling is still a bit of
mess (though this is somewhat expected).