PumiceDB write implementation / new raft backend API
This patch adds the write supplements functions which are to be for
passing KVs from the state machine back to the raft-server logic.

Added a new raft-backend-api call:  rib_sm_apply_opt().
This is called after the SM callback is run in commit-apply context.
The SM is to present all KVs it wishes to have written into the backend
- it does not write these itself!  The raft backend implementation will
execute the write itself and may encode other non-SM related metadata,
such as the last-applied raft index.