rib_entry_write() gets write supplement parameter
The raft_instance_backend rib_entry_write func pointer now accepts a
raft_net_sm_write_supplements pointer.  This pointer is optional and
will be used (at this time) only by the rockdsDB backend.

Note the write api may be (and should be) simplified in the future to
accept one set of KVs instead the raft-entry pointer and a set of KVs.
The raft-entry itself can be passed as a KV in the same set as the
supplements.