Refactoring / Re-org of pumice and raft client
During the process of building a test server for pumiceDB, it became
apparent that some code refactoring was needed to create the proper
abstractions for both the server and client libraries.

This patch introduces struct raft_net_client_user_id which is used for
holding the 'user' information of an application rather than a network
peer / client.  raft_net_client_user_id will be used to identify the
individual sub-services which are running behind a given peer.  One
example would be a niova-chunk, which is a member of a vdev.  The vdev
has a UUID whereas the chunk uses a 64-bit id.  raft_net_client_user_id
has 48-bytes of storage available for use.  Note that pumiceDB will now
use this raft_net_client_user_id as its key for the pmdb_object.