Modify raft_net_client_request_handle code
This patch changes the way that raft_net_client_request_handle are
managed to a moderate degree.  The goal being to unify the usage of the
raft_net_client_request_handle, to the extent possible, when used from
client RPC context and from raft commit context.  The main differences
between these are:
 - RPC context has the RPC request and the reply information (though a reply
may not be made immediately).
 - Raft-commit context will never have the original RPC request and the
reply information must be obtained from the state machine.  Furthermore,
the reply info may not be obtainable if Raft had undergone a leadership
change while the request was in progress.  This is because the current
leader did not take the initial write request from the client, instead,
it has learned about this request through the append entry / raft log
replication.   Log replication currently does not contain client network
information, which means a new leader may not issue a reply to a request
which it did itself receive.