More raft_client work.
Getting closer on the submit / reply processing.

Started to implement the retry framework.

Added some support for an event pipe on the raft client.  This is to be
used to schedule retries.  The method I have currently may flood the
raft leader if the retry queue gets too large.   I think this may have
to be revisited to avoid this problem.  An alternative may be to use the
timercb thread on a higher frequency than the default when retries are
pending.