Raft client rpc queuing changes
This patch starts to unify the retry and initial request submission code
paths.  I decided to go this route so that a single rate limiting source
could be maintained.

The method uses evp's - which will receive an event on app submission or
when the sendq is found to have unsent entries.