raft_net_evp_add() returns the index used
Atm, the raft_server assumes that it gets specific indices and uses
those respective index values to access the evp array.  This patch
causes the idx to be returned but does not use the idx intelligently.
Instead it just asserts that the static indices are consistent with the
assigned indices.