Adding support to enforce max number of sub apps
This patch starts to add support for keeping the heap of sub apps below
a max amount.  To do this, I've added a raft_client_instance pointer
into the raft_client_sub_app structure so that the count may be managed
from the constructor / destructor.  This code uses an atomic since the
constructor is not run inside the mutex.

There are various compilation fixes in this patch as well. However, the
code still does not yet compile..