Raft client fixes around ping handling
This patch adds some very basic back-off handling so that client pings
are not issued during every timerfd callback when a leader is not
'viable'.   A ping has been added to the viable leader as well.  This
should be done whenever the leader has been labeled as idle for some
time.