Add ctl-interface support for the raft client op cache
This patch starts to add support for placing completed operations into a
cache and exporting that cache through the ctl-interface.

It may be the case that all operations, not just recently completed,
need to be accessed in this manner.  For now, this implementation takes
a very simple approach - it merely copies the contents of the sa into a
fixed size array (which is used as an LRU).  There are no references on
the objects..