Pumice DB Client work
This patch rips up the makefile and removes the shared libraries for
pumiceDB.  This was because I was having trouble with debugging
symbols and didn't want to spend a lot of time on that issue.  The
changes here are temporary and should be removed when pumiceDB is split
out from the main niova tree.

Added PmdbObjLookup().

test/pumice_db_test_client.c now has a working ctl-interface which can
can take commands in the form of:

	APPLY input@455f648c-d378-11ea-be13-90324b2d1e89:0:0:0:0.lookup
	WHERE /pumice_db_test_client/input
	OUTFILE /pmdb.out

These commands are executed and turned into pumiceDB requests.  This
patch attempts to implement the handling for an object lookup.