Pumice_db client / raft client blocking request fix
There was a nasty circular dependency problem which would occur when a
blocking request would time out or be canceled.  The issue was that the
application callback would be issued only at the time that last
reference to the request was released.  However, the blocking thread was
holding a reference to the request!  This patch clears up the problem by
disallowing the blocking thread from timing out on its own - now, the
raft client request processor must detect the timedout request and
cancel it.  Threads blocked on canceled requests are woken in the
request's destructor context.