Numerous PMDB bug fixes while debugging 'put'
Here are a slew of fixes which have been accumulating over the past few
days.  They are the result of debugging the 'put' function.

Added a new API call, PmdbCfHandleLookup(), and modified PmdbExec() to
accept column-family names from the application layer.  These are needed
so that the application can use its own CFs.

PmdbObjPut() and PmdbObjPutNB() get a pmdb_obj_stat parameter.  This is
so that the write-sequence-number of the 'put' operation can be
provided.  The pmdb_obj_stat pointer may also be used for storing the
stat info of the object following a reply from the PMDB backend.

Fix thinko in PMDB_MAX_APP_RPC_PAYLOAD_SIZE_UDP declaration.

pmdb_object_init() - the pending term of new objects is now set to -1.
Setting it to the current term was causing some usage inconsistency
since some code aspects were using the PT to test for the object's
presence in the system.

pmdb_sm_handler_client_write() now treats -EINPROGRESS as a system
error.  This is so that the client will retry the operation after some
time without exposing any error to the application.  The bug this fixes
was the case where raft commit was pending while a retry arrived at the
leader.  Upon seeing the retry, -EINPROGRESS was returned to the
application as an 'app-error' which caused the app to complete and fail
the request.

pmdb_sm_handler() was not fully accounting for being called by 'apply'
context instead of 'write' context.  In 'apply' context, the
rncr_request pointer is expected to be NULL.

pmdb_obj_put_internal() fix a pasto where iov_len was being set to
"sizeof(sizeof(struct pmdb_msg))".

raft_net_write_supp_get() had a thinko where an array index was being
used after it was incremented.