PMDB client test fixes
This patch appears to make GETs work for a given sub-app.  Prior, there
was an issue with validation of the apps' stored values which was caused
by the incorrect value being placed into the write cmd.  The sum / total
was being used instead of the next value in the app's sequence.  This
patch adds a new raft_test_values struct which holds the values of the
last sequence and app-value to be drawn from the random state info in
the papp.

The patch also addresses some reporting issues - showing the error status of
the papp and sync status of the current operation.