PMDB more error handling fixes and add lookup to read op
raft_client_reply_try_complete() was not copying the app reply contents
if any error was found.  I switched this logic to apply only when the
reply size appears to be incorrect.

Changed some error handling in pmdb_sm_handler_client_write() so that
rcrm_app_error is replaced by pmdbrm_error.  This is in the case of a
PMDB sequence error.  Here, the raft_client reply should not be
concerned about the upper level error.

Added pmdb_object_lookup() to the pmdb_sm_handler_client_read() call so
taht the pmdb object information would be transferred back to the
client.  This is needed for PmdbGetX().