Add reh_store_time to the raft header
"store_time" records the time at which the entry was written into the
raft log.  This value is not included in the header's checksum since it
will differ across raft peers causing inconsistent CRCs.

raft_server_state_machine_apply() uses this value to determine the
commit / apply latency of the operation.  At this time, only the leader
computes and stores this information.