Raft async entry append - first draft
This patch starts work on an async write feature which will alleviate
the need for the raft thread to perform synchronous writes for entry
appends.  The motivation is to remove the large synchronous write time
and allow for multiple append-entry operation to complete within a
single backend flush operation.  Currently, only a single AE operation
is done per flush and the flushes typically take several milliseconds.

The implementation will allow for the current all-sync-writes behaviour
to be maintained should the user choose.  raft_server_instance_run() now
has a new option to enable sync_writes.

The patch here does not yet work (or even compile), it's more of a
checkpoint.  It does start to implement reporting through the
ctl-interface, adding entries for the unsynced entries which are
pending.

There's a new backend API call, rib_backend_sync(), whose job it is to
sync all pending writes into the respective backing store.  This API
call is wrapped by raft_server_backend_sync(), which samples the current
unsync'd entry header contents and tries to apply them to the sync'd
contents (if they represent the last sync operation performed).

A new thread will arrive in the next patch.  This thread causes the need
for some synchro in the raft instance structure, specifically around the
newest-entry members.  The newest sync entry may be updated from both
the raft and sync threads now.  It's important that the
newest-sync-entry does not decrease in value or become corrupted (since
it spans 2 cache lines).   The methods for obtaining members of these
structures are also protected by the mutex.  While I don't think this
was strictly necessary (small, single value reads), it seemed to provide
a more uniformity when dealing with the newest_header entries.